+++ tests/unit_tests/test_registries/test_utils.py
